+++ src/main/java/org/example/fibonacci/FibonacciRecursive.java
@@ -1,12 +1,13 @@
 package org.example.fibonacci;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public class FibonacciRecursive implements FibonacciStrategy {
     @Override
-    public BigInteger solveFibonacci(int n) {
+    public BigDecimal solveFibonacci(int n) {
         if (n <= 1) {
-            return BigInteger.valueOf(n);
+            return BigDecimal.valueOf(n);
         }
 
         return solveFibonacci(n - 1).add(solveFibonacci(n - 2));
